Complete Phase 1: disclosure, transcript persistence, call recording, STT error logging
Four remaining Phase 1 items implemented and verified with live phone calls:

- Recording disclosure: TTSSpeakFrame queued before pipeline run (append_to_context=False)
  to play legal notice at call start without polluting LLM context
- Transcript persistence: TranscriptCapture processor saves user + assistant turns to
  PostgreSQL. Two instances needed — one between STT/aggregator (user), one between
  LLM/filter (assistant) — because TranscriptionFrame is consumed by user_aggregator
- Call recording: AudioBufferProcessor captures audio, WAV created in finally block,
  uploaded to MinIO via upload_recording_async. recording_url saved on Call record
- STT error logging: on_error handler on DeepgramSTT. Hot-swap deferred to Phase 6

Bug fixes found during verification:
- timestamp_ms Integer→BigInteger (ms epoch overflows int32)
- CallResponse.started_at str→datetime (caused /calls 500)

+++ src/pipesong/models/call.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from pipesong.services.database import Base
@@ -28,5 +28,5 @@ class Transcript(Base):
     call_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("calls.id"))
     role: Mapped[str] = mapped_column(String(20))  # 'user' or 'assistant'
     content: Mapped[str] = mapped_column(Text)
-    timestamp_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    timestamp_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
